Shifted to WebGL2

+++ src/components/Canvas.js
@@ -1,4 +1,5 @@
 import React, { useState, useRef, useEffect } from "react";
+import renderGradient from "../js/gradientRenderer";
 
 var worker = new window.Worker("./gradientWorker.js");
 
@@ -11,44 +12,48 @@ const Canvas = ({ id, canvasPoints }) => {
         window.requestAnimationFrame(() => draw(imageData));
     };
     const shootPixel = () => {
-        var ctx = canvas.current.getContext("2d");
         canvas.current.width = canvas.current.offsetWidth;
         canvas.current.height = canvas.current.offsetHeight;
-        const imageData = ctx.createImageData(
-            canvas.current.width,
-            canvas.current.height
-        );
-        var imDataLength = imageData.data.length;
-        // Calling worker
-        worker.terminate();
-        // worker = new window.Worker("./gradientWorker.js", { type: "module" });
-        worker = new window.Worker("./gradientWorker.js");
-        worker.postMessage({
-            imageData: imageData,
-            points: points,
-            canvas: {
-                width: canvas.current.width,
-                height: canvas.current.height,
-            },
-        });
-        worker.onerror = (err) => {
-            console.log("error", err);
-        };
-        worker.onmessage = (e) => {
-            if (imDataLength === e.data.imageData.data.length) {
-                var t0 = Date.now();
-                draw(e.data.imageData);
-                var t1 = Date.now() - t0;
-                console.log("drawing time " + t1 + "ms");
-            }
-        };
+        if (!canvas.current.getContext("webgl2")) {
+            var ctx = canvas.current.getContext("2d");
+            const imageData = ctx.createImageData(
+                canvas.current.width,
+                canvas.current.height
+            );
+            var imDataLength = imageData.data.length;
+            // Calling worker
+            worker.terminate();
+            worker = new window.Worker("./gradientWorker.js");
+            worker.postMessage({
+                imageData: imageData,
+                points: points,
+                canvas: {
+                    width: canvas.current.width,
+                    height: canvas.current.height,
+                },
+            });
+            worker.onerror = (err) => {
+                console.log("error", err);
+            };
+            worker.onmessage = (e) => {
+                if (imDataLength === e.data.imageData.data.length) {
+                    // var t0 = Date.now();
+                    window.requestAnimationFrame(draw(e.data.imageData));
+                    // var t1 = Date.now() - t0;
+                    // console.log("drawing time " + t1 + "ms");
+                }
+            };
+        } else {
+            var t0 = Date.now();
+            renderGradient(points, canvas.current);
+            var t1 = Date.now() - t0;
+            console.log("GPU drawing time " + t1 + "ms");
+        }
     };
-
     useEffect(() => {
         setPoints(canvasPoints);
         shootPixel();
     });
-
     return <canvas id={id} ref={canvas} />;
 };
